+++ GRBobs/src/GRBobsSim.cxx
@@ -120,7 +120,7 @@ GRBobsSim::GRBobsSim(GRBobsParameters *params)
 void GRBobsSim::GetUniqueName(void *ptr, std::string & name)
 {
   std::ostringstream my_name;
-  my_name << reinterpret_cast<int> (ptr);
+  my_name << reinterpret_cast<long> (ptr);
   name = my_name.str();
   gDirectory->Delete(name.c_str());
   reinterpret_cast<TH1*> (ptr)->SetDirectory(0);
